Fix double applications of renames on features
Some GFFs contain the exon that is a child of multiple mRNAs,
which breaks the tree structure of features, and my traversal visits
some nodes more than once.

OTOH, the renaming scheme contains {mRNA -> __mRNA, exon -> mRNA}
hence the twice application leads unexpected results like exon -> __mRNA.

This patch introduces a dummy prefix defined as
RenameHandler._DUMMY_PREFIX = "__tempname__",
and it prevents collisions of a key and a value by
{mRNA -> __tempname____mRNA, exon -> __tempname__mRNA}.

The dummy prefix is removed in the second scan.